Update and rename MessageStuffGUI.java to MessageStuff.java
Message gui made visible in its class and imports added
